migrate makeStyles fo styled in login and NotFound pages
