+++ template-base-3-typescript/template/src/routes.tsx
@@ -6,7 +6,7 @@ const Main = lazy(
   () => import(/* webpackPrefetch: true */ 'components/views/main/Main'),
 );
 const NotFound = lazy(() => import('components/views/NotFound'));
-const Login = lazy(() => import('components/views/Login'));
+const Login = lazy(() => import('components/views/login/Login'));
 // [hygen] Import views
 
 export const ROUTE_PATHS = {
