+++ template-base-3/template/src/routes.js
@@ -4,7 +4,7 @@ import DefaultView from 'components/common/DefaultView';
 
 const Main = lazy(() => import(/* webpackPrefetch: true */ 'components/views/main/Main'));
 const NotFound = lazy(() => import('components/views/NotFound'));
-const Login = lazy(() => import('components/views/Login'));
+const Login = lazy(() => import('components/views/login/Login'));
 // [hygen] Import views
 
 export const ROUTE_PATHS = {
